requirements.txt
llama-cpp-python
fastapi
uvicorn
pydantic
pytest
httpx
pytest-asyncio
prometheus_client
